Refactor checkout and new
